implemented edit functionality

+++ src/main/java/com/sirmaacademy/oopexam/employeemanagementsystem/entity/Employee.java
@@ -93,8 +93,8 @@ public class Employee {
         return salary;
     }
 
-    public void setSalary(BigDecimal salary) {
-        this.salary = salary;
+    public void setSalary(double salary) {
+        this.salary = BigDecimal.valueOf(salary);
     }
 
     /**
